hugging_face/requirements.txt
gradio==6.4.0
coqui-tts==0.27.3
torch==2.9.1
scipy==1.16.3
transformers==4.57.3
huggingface_hub